Descargar comprobante de pago

+++ src/main/kotlin/com/estonianport/agendaza/model/Pago.kt
@@ -23,7 +23,7 @@ open class Pago (
     val id: Long,
 
     @Column
-    val monto : Int,
+    val monto : Double,
 
     @PrimaryKeyJoinColumn
     @Enumerated(EnumType.STRING)
